Refactor: extract lib/zendesk_api/search.rb
From lib/zendesk_api/resources.rb
